Use own component for Route layer (drawing features, gpx, document)
