+++ src/components/yeti/map-layers/layer.js
@@ -1,0 +1,11 @@
+export default {
+  computed: {
+    map() {
+      // store parent map
+      return this.$parent.map;
+    },
+    view() {
+      return this.map.getView();
+    },
+  },
+};
